+++ src/parlant/adapters/nlp/openai.py
@@ -34,7 +34,7 @@ from pydantic import ValidationError
 import tiktoken
 
 from parlant.adapters.nlp.common import normalize_json_output
-from parlant.core.engines.alpha.guideline_proposer import GuidelinePropositionsSchema
+from parlant.core.engines.alpha.tool_caller import ToolCallInferenceSchema
 from parlant.core.logging import Logger
 from parlant.core.nlp.policies import policy, retry
 from parlant.core.nlp.tokenization import EstimatingTokenizer
@@ -48,9 +48,6 @@ from parlant.core.nlp.generation import (
     UsageInfo,
 )
 from parlant.core.nlp.moderation import ModerationCheck, ModerationService, ModerationTag
-from parlant.core.services.indexing.guideline_connection_proposer import (
-    GuidelineConnectionPropositionsSchema,
-)
 
 
 class OpenAIEstimatingTokenizer(EstimatingTokenizer):
@@ -200,9 +197,9 @@ class OpenAISchematicGenerator(SchematicGenerator[T]):
                         ),
                     ),
                 )
-            except ValidationError:
+            except ValidationError as e:
                 self._logger.error(
-                    f"JSON content returned by {self.model_name} does not match expected schema:\n{raw_content}"
+                    f"Error: {e.json(indent=2)}\nJSON content returned by {self.model_name} does not match expected schema:\n{raw_content}"
                 )
                 raise
 
@@ -381,9 +378,9 @@ class OpenAIService(NLPService):
 
     @override
     async def get_schematic_generator(self, t: type[T]) -> OpenAISchematicGenerator[T]:
-        if t == GuidelineConnectionPropositionsSchema or t == GuidelinePropositionsSchema:
-            return GPT_4o_24_08_06[t](self._logger)  # type: ignore
-        return GPT_4o[t](self._logger)  # type: ignore
+        if t == ToolCallInferenceSchema:
+            return GPT_4o[t](self._logger)  # type: ignore
+        return GPT_4o_24_08_06[t](self._logger)  # type: ignore
 
     @override
     async def get_embedder(self) -> Embedder: